Fix filtering of meetings in list view
This also moves the sorting to the selector
