+++ app/routes/meetings/MeetingListRoute.js
@@ -4,12 +4,12 @@ import { connect } from 'react-redux';
 import { dispatched } from '@webkom/react-prepare';
 import { fetchAll } from 'app/actions/MeetingActions';
 import { LoginPage } from 'app/components/LoginForm';
-import { selectMeetings } from 'app/reducers/meetings';
+import { selectGroupedMeetings } from 'app/reducers/meetings';
 import replaceUnlessLoggedIn from 'app/utils/replaceUnlessLoggedIn';
 import MeetingList from './components/MeetingList';
 
 const mapStateToProps = (state, props) => ({
-  meetings: selectMeetings(state),
+  meetingSections: selectGroupedMeetings(state),
   currentUser: props.currentUser,
   loading: state.meetings.fetching
 });
